refactor(responses): split content, tool, and text-format parsing out of the request parser (split S07 L1/4)
